Preparation for 0.1.0-beta.1 (64)

+++ scirs2-integrate/src/symplectic/leapfrog.rs
@@ -12,6 +12,7 @@ use crate::common::IntegrateFloat;
 use crate::error::IntegrateResult;
 use crate::symplectic::{HamiltonianFn, SymplecticIntegrator};
 use ndarray::Array1;
+use std::f64::consts::PI;
 use std::marker::PhantomData;
 
 /// Störmer-Verlet Method (also known as Leapfrog)
@@ -154,7 +155,7 @@ mod tests {
 
         // Integrate for a full period
         let period = 2.0 * PI;
-        let steps = (period / dt).round() as usize;
+        let steps = ((period / dt) as f64).round() as usize;
 
         let integrator = StormerVerlet::new();
         let mut q = q0.clone();
@@ -192,7 +193,7 @@ mod tests {
 
         // Integrate for one period
         let period = 2.0 * PI;
-        let steps = (period / dt).round() as usize;
+        let steps = ((period / dt) as f64).round() as usize;
 
         // Test StörmerVerlet
         let mut q1 = q0.clone();
